+++ coles_vs_woolies/examples.py
@@ -7,6 +7,7 @@ from rich import box
 from rich.console import Console
 from rich.table import Table
 
+from coles_vs_woolies.search.similarity import jaccard_similarity
 from coles_vs_woolies.search.types import Merchant, Product, ProductOffers
 
 _console = Console()
@@ -23,7 +24,8 @@ def compare_offers(product_offers: ProductOffers):
         for i, _product in enumerate(products):
             txt_colour = 'green' if is_sales and i in cheapest_product_idx else 'grey50'
             # txt_colour = None if not i else 'grey50'
-            _console.print(f'  {_product.merchant.upper()}: {_product}', style=txt_colour)
+            similarity = jaccard_similarity(name, _product.display_name)
+            _console.print(f'  {_product.merchant.upper()}: {_product} | {similarity=:.2f}', style=txt_colour)
     _console.print('\n')